# openapi.yaml
openapi: 3.1.0
info:
  title: Affilite Mix API
  version: "1"
  description: "Multi-tenant affiliate marketing platform API. Routes marked x-internal: true are not part of the public consumer contract."
  contact:
    name: Affilite Mix
  license:
    name: Proprietary
servers:
  - url: "https://compareai.site"
    description: Production (compareai.site)
  - url: "https://staging.compareai.site"
    description: Staging
paths:
...
  /api/newsletter/unsubscribe:
    get:
      summary: GET /api/newsletter/unsubscribe (site) — auth=token
      operationId: getApiNewsletterUnsubscribe
      tags:
        - Newsletter
      x-auth: token
      x-adminRequired: false
      x-tenantScope: site
      x-rateLimit: true
      x-csrf: false
      x-requestSchema: "{ token: string }"
      x-responseSchema: Ok
      x-notes: null
      security:
        - bearerAuth: []
      responses:
        200:
          description: Successful response (Ok)
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
          content:
            application/json:
              schema:
                type: object
        401:
          description: Unauthorized
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
        429:
          description: Rate limited
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
    post:
      summary: POST /api/newsletter/unsubscribe (site) — auth=token
      operationId: postApiNewsletterUnsubscribe
      tags:
        - Newsletter
      x-auth: token
      x-adminRequired: false
      x-tenantScope: site
      x-rateLimit: true
      x-csrf: false
      x-requestSchema: "{ token: string }"
      x-responseSchema: Ok
      x-notes: null
      security:
        - bearerAuth: []
      responses:
        200:
          description: Successful response (Ok)
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
          content:
            application/json:
              schema:
                type: object
        401:
          description: Unauthorized
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
        429:
          description: Rate limited
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
...
components:
...
  headers:
    API-Version:
      description: Numeric major version of the served API contract.
      schema:
        type: string
        pattern: "^[1-9][0-9]*$"
        example: "1"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...